[PM-40679] Resolve circular dep
